requirements.txt
tqdm~=4.60.0
requests~=2.25.1
beautifulsoup4~=4.9.3
pypandoc~=1.7.2
python-dotenv~=0.16.0